requirements.txt
aiogram==3.4.1
fastapi
uvicorn
sqlalchemy
aiosqlite
apscheduler
python-dotenv
